pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-research-intelligence-platform"
version = "0.1.0"
description = "Hybrid AI system for research document ingestion, processing, and RAG-powered retrieval"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.9",
    # Document processing
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    "beautifulsoup4>=4.12.0",
    "markdown-it-py>=3.0.0",
    "chardet>=5.2.0",
    # Embeddings & vector store
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    # Pipeline & utilities
    "tiktoken>=0.6.0",
    "python-dateutil>=2.8.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
